Refactoring - Remove some schemes
