+++ api/v1/endpoints/pig.py
@@ -2,14 +2,14 @@ from typing import List, Optional
 
 from fastapi import APIRouter, status
 
-from schemas.requests.pig import PIGCreationRequest, PIGUpdateRequest
-from schemas.responses.pig import PIGDeleteResponse, PIGResponse
+from models.pig import PIGModel
+from schemas.pig import PIGCreationRequest, PIGDeleteResponse, PIGUpdateRequest
 from services.pig import PIGService
 
 router = APIRouter()
 
 
-@router.post("/", status_code=status.HTTP_201_CREATED, response_model=PIGResponse)
+@router.post("/", status_code=status.HTTP_201_CREATED, response_model=PIGModel)
 def create_pig(pig_body: PIGCreationRequest):
     pig_record = PIGService().create(
         serial_number=pig_body.pig_number,
@@ -17,49 +17,22 @@ def create_pig(pig_body: PIGCreationRequest):
         company=pig_body.company_id,
         description=pig_body.description,
     )
-    return PIGResponse(
-        id=pig_record.id,
-        pig_number=pig_record.pig_number,
-        name=pig_record.name,
-        company_id=pig_record.company_id,
-        description=pig_record.description,
-    )
+    return pig_record
 
 
-@router.get("/", status_code=status.HTTP_200_OK, response_model=List[PIGResponse])
+@router.get("/", status_code=status.HTTP_200_OK, response_model=List[PIGModel])
 def get_pigs(company_id: str):
     pig_records = PIGService().get_all_by_company(company=company_id)
-    pigs_reponse = []
-    for pig_record in pig_records:
-        pigs_reponse.append(
-            PIGResponse(
-                id=pig_record.id,
-                pig_number=pig_record.pig_number,
-                name=pig_record.name,
-                company_id=pig_record.company_id,
-                description=pig_record.description,
-            )
-        )
-
-    return pigs_reponse
+    return pig_records
 
 
-@router.get(
-    "/{pig_id}", status_code=status.HTTP_200_OK, response_model=Optional[PIGResponse]
-)
+@router.get("/{pig_id}", status_code=status.HTTP_200_OK, response_model=PIGModel)
 def get_pig(pig_id: str):
     pig_record = PIGService().get_by_id(pig_id=pig_id)
-
-    return PIGResponse(
-        id=pig_record.id,
-        pig_number=pig_record.pig_number,
-        name=pig_record.name,
-        company_id=pig_record.company_id,
-        description=pig_record.description,
-    )
+    return pig_record
 
 
-@router.put("/{pig_id}", status_code=status.HTTP_200_OK, response_model=PIGResponse)
+@router.put("/{pig_id}", status_code=status.HTTP_200_OK, response_model=PIGModel)
 def update_pig(pig_id: str, pig_body: PIGUpdateRequest):
     pig_record = PIGService().update(
         pig_id=pig_id,
@@ -67,14 +40,7 @@ def update_pig(pig_id: str, pig_body: PIGUpdateRequest):
         company=pig_body.company_id,
         description=pig_body.description,
     )
-
-    return PIGResponse(
-        id=pig_record.id,
-        pig_number=pig_record.pig_number,
-        name=pig_record.name,
-        company_id=pig_record.company_id,
-        description=pig_record.description,
-    )
+    return pig_record
 
 
 @router.delete(
