+++ schemas/healthcheck.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
 
+################# Responses #################
 class HealthCheckResponse(BaseModel):
     status: str
