+++ schemas/pig.py
@@ -3,6 +3,7 @@ from typing import Optional
 from pydantic import BaseModel
 
 
+################# Requests #################
 class PIGCreationRequest(BaseModel):
     name: str
     pig_number: str
@@ -14,3 +15,9 @@ class PIGUpdateRequest(BaseModel):
     name: Optional[str]
     company_id: Optional[str]
     description: Optional[str]
+
+
+################# Responses #################
+class PIGDeleteResponse(BaseModel):
+    id: str
+    status = "deleted"
